state管理をLayoutに変更
